sol with bad TypeScript
